feat: use Network framework for WebSocket
Remove NWWebSocket and implement a concurrent wrapper over
Network.framework instead. This allows us to close the WebSocket with a
certain close code (behavior which the library did not seem to implement
correctly), and use Combine to handle event and message plumbing.

Unfortunately, because everything is using Swift Concurrency now, there
doesn't seem to be an easy way to disconnect the client whilst blocking
on that operation, which is vital for deinit and
applicationWillTerminate. This will have to be solved later.