refactor: adopt a concrete generic json repr
Having the main event data of a gateway packet have the Any type is not
only dangerous, but unpleasant to use. Adopt a generic JSON type
instead, which is easier to query and manipulate.

We could've possibly used the generic JSON representation from
RichJSONParser, but I found GenericJSON to be better. We should write
some interoperability between the two so we can use just a single JSON
type.