feat: maintain position when scrolling up
If we naively prepend new messages to the collection view, the scroll
position remains the same after adding new messages (e.g. if we are 10
points from the top, the scroll view will remain 10 points scrolled from
the top after fetching).

This is unintuitive because it means that as the user scrolls upwards,
all of the messages onscreen will be replaced with the older messages.
Instead, what is expected is that the messages are prepended without
clobbering the scroll position (the messages invisibly appear offscreen
for the user to scroll into).

We can implement this by saving the scroll offset and restoring it
(adding the difference in content height) after the messages have been
restored. Not only does this feel much better, but it lets us remove the
throttling behavior since we don't get trapped in a loop of loading
older messages constantly.