refactor: switch to codable + SwiftyJSON
Prior to this commit, Basilisk handled all JSON interchange with 3
libraries: GenericJSON, RichJSONParser, and FineJSON.

1. GenericJSON was used to ease nested object access with subscripts and
   optional chaining, because manual casting is noisy and error prone.
2. FineJSON was used to achieve deeper control in how we serialized
   JSON; e.g., how nulls, missing fields, and the ordering of object
   keys were handled.
3. RichJSONParser contains FineJSON's underlying JSON representation.
   This was used to manually construct large JSON objects, which were
   then used with FineJSONEncoder.

Remove all of these libraries and replace them with Codable and
SwiftyJSON. This is done for the following reasons:

- Decoding of data models previously made heavy use of the force unwrap
  (`!`) operator, which is unacceptable if we want to be able to
  introduce features to help debug and inspect gateway packets within
  the app itself. In the future, we need to gracefully recover from
  situations in which the structure of a gateway packet is not what we
  were expecting, and it should be easy to see what went wrong.
- While nested object access is made easier with subscripts, it's
  noisier than simply using Decoder, which also allows us to move the
  details of how certain types are decoded into their `init(from:)`
  method, and away from usage sites. (For example, data models that
  contain snowflakes no longer have to know that snowflakes are
  represented as strings of integers. We just call a method to "decode
  the snowflake", and the right thing happens.)
- While the order of object keys is important if we want our outgoing
  data to remain bit-for-bit identical to the official client, it's
  likely that Discord doesn't actually care. We can do a workaround for
  this late to regain parity; the benefits largely outweigh this single
  disadvantage at the moment.
- SwiftyJSON has JSON literals, eliminating the need for enum cases all
  over the code when we need to create a JSON object.
- Implementing Codable means that we get serialization, too! This could
  be useful for caching (!) and outbound packets.

Unfortunately, due to use of Codable, we can no longer pass a generic
"JSON" blob everywhere anymore. We either decode to that blob, or to a
class implementing Decodable. Once you choose, though, it's somewhat
difficult to change representations without wastefully encoding
something that we just decoded. To remedy this, make GatewayPacket
generic. The type parameter becomes the event data ("d" key in the
packet).

Then, when inspecting packets for processing, we can use
GatewayPacket<JSON>. This represents a gateway packet with a (currently)
unknown event data shape. This is further wrapped within a type named
AnyGatewayPacket, which additionally contains the raw packet data, and
is used in the gateway inspector. A convenience method named reparse
allows the event data to be parsed into a real type implementing
Decodable, which is what the raw data is needed for.

(This _does_ mean that we parse the event data an extra time in most
cases, but we might be able to fix this later by not parsing the event
data when we aren't sure what shape it has, instead of always parsing it
into a generic blob first. However, this may be impossible in some
cases.)

When given an AnyGatewayPacket, we can inspect the event name and
reparse it succinctly like so:

    // We have _some_ gateway packet.
    if anyGatewayPacket.eventName == "MESSAGE_CREATE" {
      // OK, we know it's a MESSAGE_CREATE. Reparse the event data into
      // a message model.
      let message: Message = anyGatewayPacket.reparse()
      print(message.content)
    }

Easy.

Furthermore, this commit adds outgoing packet support to the gateway
inspector and updates the default disguise to match the latest build of
Discord PTB.