add base support for direct messages
Support for leaving group DM channels, editing them, etc. is not yet
present. Also, because we are missing a cache, we are only able to
display the channel ID. After adding one, we can have the proper UI for
this.

Including an assortment of other chat-related fixes:

- If the message history happens to end on a 50 message boundary, then
  properly trip the "no more messages" exhaustion flag.
- Also trip the exhaustion flag if the channel had less than 50 messages
  to begin with.

And Serpent changes:

- Rename Channel to GuildChannel to better reflect the class's intended
  usage. Added DMChannel and GroupDMChannel to support direct message
  channels, and add PrivateChannel as an enumeration that encapsulates
  both type of DM channel. This is used to parse the private_channels
  field in the READY payload.
- Introduce a JSONDecoder subclass that uses ISO 8601 date parsing by
  default, and use it when parsing JSON in the right contexts.