add a cache
Add a Cache actor to client instances. Because Cache instances are
actors, all access to them are gated behind suspension points (await),
which can be a pain. However, this ensures that access from other
threads is completely safe assuming Swift Concurrency is used. Having
the Cache live separately also means that it's going to be easier to
serialize and restore it.

Because context switching while the outline view is requesting cells is
a nonstarter, the chat view controller maintains its own copies of the
known current user, known private channels, known guilds, and any known
private participants (users who the current user has private channels
with, or users who are in group direct message channels with the current
user). This may seem unintuitive as the cache is the preferred single
source of truth, but trying to access the cache while configuring cells
is worse.

Included in this commit is a myriad of other changes:

- Fix Snowflakes having a broken Comparable conformance.
- Make the direct message list take advantage of the cache by correctly
  resolving other users' names and avatars.
- Properly sort the private channel list in the same way that the main
  client does.
- When setting up Combine subscribers, use .store(in: &set) to reduce
  the amount of instance variables necessary. This is a neat feature!
- Remove extraneous `?`s from URLs when no query is provided.
- Make DMChannel, GroupDMChannel, and User conform to Hashable.
- Change GroupDMChannel.recipients to a Set<Ref<User>>.
- Lots of polish to make working with Ref<T> easier. Notably, it now
  conforms to Hashable (just like Snowflake), and you can easily convert
  objects that conform to Identifiable and Snowflakes into Refs. You can
  also compare Refs now.
- Nuke current unit tests and replace them with a single one that tests
  private channel sorting.
- Separate the parsable READY packet into its own file.
- Make the chat view controller use snowflakes instead of UInt64
  instances.
- Remove traces of Glare.
- Bump the minimum Swift version to 5.7.
- Change the navigator outline view's row style to medium.