improve image loading behaviors
Use the same code path for loading external images every time, and be
sure to downscale the downloaded image to fit into the image view. Both
the original and resized image should be cached.

This means that a loading indicator appears for all externally loaded
images now. Behavior was added to constrain the view to the image view's
size if it's small enough, which avoid clipping and therefore looks
better.