lay the groundwork for multiple tabs and windows
Multiple tabs and windows are now supported. To facilitate this,
decouple the connection initiation logic from the chat view controller,
and have the application delegate manage chat window controllers.

The concept of a session was also introduced, which couples an Account
(account information, e.g. token and API base URL) and an active Client
together. The application delegate is also responsible for keeping
sessions alive. Chat view controllers can be associated with a session,
which instructs it to act upon and display information from that
session. This way, chat windows (and tabs, since tabs are effectively
just windows) serve as views into sessions, where the ultimate truth is
stored. Assuming all of our logic is sound, then all of the windows tied
to a particular session should never desync, aside from window-local
state such as the currently focused channel.

To make this possible, the setup logic surrounding state change handlers
(which we use Combine for) have been consolidated and normalized. Menu
items to create tabs have been added.

A consequence of this change is that Account base URL and gateway URLs
are now finally respected.