measure accessory heights more accurately
This commit includes a myriad of changes that fixes scrolling in the
messages view when message accessories are being displayed (for now, the
only accessories supported are image attachments). Janky scrolling was
caused by incorrect height measurements; for some reason, the incorrect
fitting size is returned for unified message rows when the content stack
view contains accessories. The size fails to encompass the fitting sizes
of the inner stack view, even though it is a direct subview. (The
content label was also being measured incorrectly--see below.)

For now, give a rough estimate of the proper height of the message row
when it contains accessories that necessitate a crude height
calculation. In the future, we should figure out why the fitting size of
the row is too small, so we can have a single code path that doesn't
rely on hardcoded constants.

Accessories shrinking themselves when the window is resized and not
growing was also fixed by removing the width and height constraints,
which were fixed to the clamped attachment dimensions. This is something
that is best handled by the intrinsic content size constraints, which
are steadfast and let the views size themselves back up when the window
gets bigger again.

And finally, because our message sizing template is a view floating in
space and is never actually rendered, we need to set a preferred max
layout width for the content label. Here we're also using hardcoded
constants, but this calculates a proper width in the event of wrapping.

With all of these improvements, there's much less jank when it comes to
scrolling now.

Other changes:

- Only bother setting up accessories for image attachments. We don't
  support video attachments at the moment.
- Invalidate our message height cache when the table view changes width.
  We then rely on the internal height cache of the table view, which
  should be OK for now. Not sure how else to do this.
- Add a lot of debugging functionality gated behind a default
  (BSLKMessageRowHeightDebugging).