don't shrink the template view's attachments
Normally, messages' attached images have low content resistance
priorities to allow them to shrink when the window is downsized.
However, when this happens to the template view, it automatically
resizes itself to an incredibly small frame that is not representative
of the table view row's true height when added to an actual view
hierarchy.

I'm not entirely sure why this happens, but it's likely caused by
constraints added to the view by the NSTableView. When the cached height
and actual height mismatch, scrolling starts to jump around.

To fix this, only set the content resistance priority to low when we're
not performing measurements. Otherwise, set it to required to simulate a
normal superview size.