migrate away from diffable data source
I did this to see if it would fix scrolling being weird since macOS 13.0
(Ventura), but now I'm not sure that changes in the frameworks are to
blame. Regardless, I prefer how this version of the code works, so I'm
keeping it now.

Erratic scrolling behaviors can be fixed by manually smashing the row
height cache (this has been fixed in macOS 13.1). This fix is included
in this commit.