render replies
Render replies to messages. The replying-to message's author is
displayed inline next to the replying message's author, separated with
"replied to". If the author is replying to themselves, then it will
instead say "continues". The avatar of the replying-to author is also
displayed before the username.

If the replying-to and reply message are adjacent, then the reply will
draw a small, upwards-pointing arrow in the header and omit the
referenced message preview, which I think is cute.

Replies will always start a new message header. I might look into
relaxing this behavior in the future (perhaps even displaying the avatar
next to the arrow within the `ReferencedMessagePreviewView` whenever
this happens), but that may be a bit involved.

SwiftUI is used for the referenced message preview and now, for all
message author names. This may be a bit risky, but it did make the
implementation for these features much faster and easier than it
would've been if standard AppKit was used. I may want to move the entire
message row view to SwiftUI, but that's even more scary.

The referenced message preview naively truncates the message content to
50 characters, and does not render markup (because we don't support
markup at all just yet). However, it does show a little paperclip
symbol if the replying-to message has an attachment, and displays
"Attachment" in italics in lieu of the message content if there is
none.

This patch additionally includes facilities for easier Asset
construction for the sake of SwiftUI previews.