add replying
Add support for replying to messages through a menu on the table view,
and also by a set of bespoke quick selection keyboard shortcuts.

To show the current message we're replying to, a new view has been added
in the stack view between the table view and the box that contains the
input field, called the "message field's accessories view". I think this
name is too long and clunky, so we should come up with another one soon.
This view animates open and closed in a nice way, although the adjacent
scroll view doesn't react in response. This is a bit tricky to fix.

The inner insets of the scroll view have been removed for now, as they
caused some clipping problems with the aforementioned animations. We'll
have to find another way to apply some inner padding. It'll probably be
better to apply it to the table view itself, perhaps through a fake cell
of some kind. Regardless, it was causing issues.

The actual contents of the view are written in SwiftUI, which I think is
an appropriate place for more SwiftUI to be. I believe the design is
good for now.

A bespoke quick selection mechanism has been added to make replying
faster. A local event monitor is installed in the application delegate,
which monitors for key presses on the up and down arrow keys (when shift
is also held down). This lets the user control an "outline floater"
which can be used to select a message to reply to via the keyboard. The
interaction ends when the shift key is released. I think this is a
pleasant advanced feature that accelerates an incredibly common action.

Moving the outline plays a sound (the volume and whether the sound is
played can be controlled through a user default), and the mechanism is
resistant to new messages being received in the current channel in the
middle of selection.

Default values have also been added to the type-safe defaults
implementation.